+++ search-films/src/components/Logo/index.jsx
@@ -1,0 +1,9 @@
+import './styles/index.css';
+
+function Logo() {
+	return (
+		<img className="logo" src='/logo-icon.svg' alt='Logo'/>
+	);
+}
+
+export default Logo;